Give warnings about dangerous casts in Maintainer mode (also for gcc)
* Fix some compiler warnings in aterm code

git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@9080 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/aterm/include/aterm/memory.h
@@ -8,7 +8,7 @@ namespace aterm
 
 static const size_t MAX_BLOCKS_PER_SIZE = 1024;
 
-static const size_t MAX_INLINE_ARITY = (1 << ARITY_BITS)-2;
+static const size_t MAX_INLINE_ARITY = ((size_t)1 << ARITY_BITS)-(size_t)2;
 //#define MAX_INLINE_ARITY    ((1 << ARITY_BITS)-2)
 
   /* To change the block size, modify BLOCK_SHIFT only! */
